console-auto/requirements.txt
# app
requests
django-environ

# utility
setuptools-scm

# opentelemetry auto instrumentation
opentelemetry-distro[otlp]

# opentelemetry instrumentation
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-urllib3
